refactor(workshops): refactor directory for workshopmanagment
