+++ src/components/ComonProfil/WorkshopManagment/index.js
@@ -1,0 +1,1 @@
+export * from './WorkshopManagment';
